Split function for compilation of parameters
